+++ JavaServer/calzone/src/main/java/com/vub/repository/InstitutionRepository.java
@@ -11,6 +11,6 @@ import com.vub.model.Institution;
  *
  */
 @Repository
-public interface InstitutionRepository extends JpaRepository<Institution, Long> {
+public interface InstitutionRepository extends JpaRepository<Institution, Integer> {
 
 }
